refactor(server): move the rls where evaluator into its own module
`rls/middleware.ts` was 2,035 lines and 400 of them were a self-contained pure evaluator that has
to be read against `@lunora/shard-engine`'s `where-sql.ts` to be reviewed at all — the two are
parallel implementations of one predicate language, and their agreement is a security property, not
a nicety.

Moved verbatim to `rls/where-match.ts`, which states that contract in its header: what the twin is
for, why it exists twice, and where the pinned truth table lives. `containsRelationPredicate` and
`matchesWhere` are the only two names the middleware used; `matchesWhere` stays re-exported from
`./middleware` so the in-process harness and the sibling middlewares keep one import site.

No behaviour change — the only edit to a moved line is a JSDoc `{@link}` that no longer resolves
across the file boundary.
